+++ src/stories/Molecules/Header.stories.ts
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from '@storybook/react-vite';
 
 import { fn } from 'storybook/test';
 
-import { Header } from '../components/Header';
+import { Header } from '../../components/Molecules/Header/Header';
 
 const meta = {
   title: 'Example/Header',
