+++ src/stories/Page.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Header } from '../components/Header';
+import { Header } from '../components/Molecules/Header/Header';
 import './page.css';
 
 type User = {
